truly random names using uuid (#594)
customers pointed out that sometimes the random names were being
repeated. This happened because they were setting seeds in their code,
which would cause the random generator to reset and return the same
name.

UUIDs are not affected by the seed in the same way (tbh not sure how
thats possible), but this prevents the issue from occuring. We tac on 5
characters from a uuid to make sure the name is random in the case of a
seed being set