Fix/Create out logdir before instantiating logger (#736)
https://github.com/rungalileo/feedback/issues/2

Shortcut:
https://app.shortcut.com/galileo/story/7053/dq-bug-filenotfounderror-errno-2-no-such-file-or-directory-root-galileo-out-none-out-log

This error was caused by an issue surfaced by the new major version
mismatch, wherein we log a warning to stdout before the local config is
assigned a run id. To fix, we create the dir `.galileo/out` before
instantiating the stdout logger in case it doesn't exist yet.