Feature/auto tokenizer (#748)
***What existing issue does this pull request close?***

This is a very small PR to address this shortcut
[ticket](https://app.shortcut.com/galileo/story/7664/dq-improve-tokenization-in-dq-auto)

In `dq.auto` we introduce a small change to the `tokenization` scheme to
improve the padding efficiency. Rather than padding every sequence to
the user specified `max_length`, we delay padding until the creation of
batches. This allows us to dynamically pad to just the longest sequence
in the batch.

This is a very easy change with `hf.Trainer`. By default if we *do not*
specify a `DataCollator` but include our `tokenizer`, `hf.Trainer` uses
the `DataCollatorWithPadding`. This Collator is a pretty simple wrapper
around `tokenizer.pad(...)`, which does exactly what we want!

***Demonstration***

As a little reference, here is behind the scenes what
`DataCollatorWithPadding` does:
```
    def __call__(self, features: List[Dict[str, Any]]) -> Dict[str, Any]:
        batch = self.tokenizer.pad(
            features,
            padding=self.padding,
            max_length=self.max_length,
            pad_to_multiple_of=self.pad_to_multiple_of,
            return_tensors=self.return_tensors,
        )
        if "label" in batch:
            batch["labels"] = batch["label"]
            del batch["label"]
        if "label_ids" in batch:
            batch["labels"] = batch["label_ids"]
            del batch["label_ids"]
        return batch
```
where `padding=True` be default meaning we pad to the longest sequence
in the batch. Also it converts 'label' to 'labels', which is what the
model expects!

---------

Co-authored-by: Bogdan Gheorghe <112427971+bogdan-galileo@users.noreply.github.com>