Feature/generation (#746)
* [x] I have added tests to `tests` to cover my changes.

***What existing issue does this pull request close?***

This PR adds generation support for seq2seq models - addressing this
shortcut
[ticket](https://app.shortcut.com/galileo/story/7661/dq-generation-support-for-encoder-decoder-models).

**Usage**
To support generation in dq, we have started by fully controlling
generation internally, guided by the users model and specified
generation parameters. We introduce a new `seq2seq.hf` integration, with
a watch function to capture:
1. User model
2. GenerationConfig - which is the standard for setting generation
parameters with hf
3. Whether the user wants to generate on the training split - Note that
this is quite costly.

In code this looks like:
```
model = HF_Model()  // Model with .generate() function
generation_config = GenerationConfig(
    max_new_tokens=15,
    do_sample=True,
    temperature=0.7,
    ....
)

watch(
    model,
    generation_config,
    generate_training_data=False
)
```

When the user calls `.finish()`, behind the scenes we use the user model
to generate over the logged data.

**Implementation Details**
Behind the scenes we currently generate over each sample sequentially
(i.e. we don't batch process). We have introduced a new key function in
`utils/seq2seq.py`: `add_generated_output_to_df`. This is a vaexified
function (doesn't materialize the full df into memory) that goes sample
by sample to generate the model output.

Here are some notes on the generation process:
1. We leverage the user tokenizer to tokenize the input data and feed
this to the model's `.generate()` function. In doing so we keep the
model on the GPU
2. From `.generate()` we only capture the generated ids. To get
generation log_probs, we run the sample, with generated labels, through
the model to get prediction (based on the generation) logits. This gives
us a lot more flexibility to handle different generation schemas and
frees us from the hidden nature of the `.generate()` function.
3. With the logits and generated ids, we can pretty much treat
everything like we are logging data + model outputs as in training!

*Special Note*: To vaexify the generation function - we have to use a
pyarrow trick to have a vaex register function return values for
multiple columns. Specifically, we take advantage of the
`pa.StructArray` and `pa.StructArray.from_arrays`. High level this lets
us define basically names 'columns' within a single struct array that we
can assign to a single vaex column. Luckily, vaex then supports a
`df.struct.flatten` operation to separate the columns in our
`StructArray`. **Note** that we take inspiration from a similar
implementation in the API
[here](https://github.com/rungalileo/api/blob/b69598e1ff1c6bdd8dcdadad9167c4ea198af9ff/api/utils/ner.py#L94).
 
***Open Questions***
@elboy3 One question is whether we want to make a more formal structure
for generation rather than having `generation` just as functions in the
`utils`. I don't really know the way to think through this, but for
example we have `data_loggers` and `model_loggers` - maybe we could have
an equivalent `generation_logger`?

It just feels a bit odd having a bunch of important logic being lumped
into the `utils` section.

***How are these changes tested?***

See test in:
- `tests/loggers/test_seq2seq.py`
- `tests/utils/test_seq2seq_utils.py`

---------

Co-authored-by: Elliott <echartock3@gmail.com>
Co-authored-by: Elliott <elliott@rungalileo.io>