fix: generation without vaex (#754)
* [x] I have added tests to `tests` to cover my changes.
* [ ] I have updated `docs/`, if necessary.
* [ ] I have updated the `README.md`, if necessary.

***What existing issue does this pull request close?***

This PR address this
[ticket](https://app.shortcut.com/galileo/story/7910)

***Bug Description***
Currently generation works with small dataset sizes, but fails when we
try over > 500 samples. We get some very strange error related to the
`StructArray` that we have to use. When I talked to Ben about this, he
mentioned that while possible this vaex `StructArray` design was fairly
temperamental and error prone - which it seems a might be.

***Proposed Solution***
The primary reason to use a vaex register function is to avoid bringing
large amounts of data into memory through lazy evaluation. However, as
addressed by a previous issue, we end up effectively forcing the
computation of the `generated_data` to avoid lazy recomputation.
Therefore, we are not really using the vaex register function correctly.

With this in mind, we propose switching to a more manual batched
processing of the generation job. This will give us more control over
generation, especially around potential errors. We basically move to a
manual batched approach using vaex's `.evaluate_iterator()`.

We introduce a new DataClass type called `BatchGenerationData`. This
data class stores all the information needed for generation. Moreover,
for batched processing we allow for initializing an empty central
`generation_data` object, followed by easy *in-place* extending with a
new batch of generation data. See this example:
```
generation_data = BatchGenerationData.empty_init()
for _ in range(num_batches):
  batch_generated_data = ...
  # Add new data to central data store
  generation_data.extend_from(batch_generated_data)
...
```

**Note**: We favor in-place adding of data for memory and performance
concerns.

---------

Co-authored-by: Elliott <echartock3@gmail.com>
Co-authored-by: Elliott <elliott@rungalileo.io>