Silencing tqdm during generation (#757)
* [ ] I have added tests to `tests` to cover my changes.
* [ ] I have updated `docs/`, if necessary.
* [ ] I have updated the `README.md`, if necessary.

***Description***
This is a very small PR addressing a bug/annoyance when generating in
`dq.finish()`. Essentially, during generation we rely no the function
`align_tokens_to_character_spans` to create token alignment data. We do
this at a per-sample level, since that is how we do generation. However,
the `align_tokens_to_character_spans` function was initially developed
as a batch processing methods. And with this, we use tqdm.

This becomes annoying when processing at the sample level because we
constantly display `tqdm` for just a single iteration loop. See the
demonstration bellow.

***Demonstration***

Constantly re-running `tqdm` with a single sample leads to this annoying
flickering on the DQ side.


https://github.com/rungalileo/dataquality/assets/29869344/88536639-b9e1-4071-a781-74f147fdf18e


***Solution***

We basically just turn of `tqdm` when generation. Here is the result!


https://github.com/rungalileo/dataquality/assets/29869344/94d36180-7e23-492a-b85e-ac7e8e414209