chore: add support for Tokenizer (#797)
Adding support for tokenizers of type Tokenizer (to support Cohere)
https://app.shortcut.com/galileo/story/8963

The class `from tokenizers import Tokenizer` is a performant
implementation of lots of popular tokenizers
https://github.com/huggingface/tokenizers. It is implemented in Rust
with bindings in Python.

The class `from transformers import PreTrainedTokenizerFast` seems more
popular as it integrates easier with transformers and many trained
tokenizers have this type
https://huggingface.co/docs/transformers/v4.33.2/en/main_classes/tokenizer#transformers.PreTrainedTokenizerFast.

We first implemented support for PreTrainedTokenizerFast and require
support for Tokenizer (in particular for Cohere). It turns out a
Tokenizer can be converted into a PreTrainedTokenizerFast, so that's
what we'll use for now since it allows minimal code change. We can think
about a more general solution if we'll need to support more general
tokenizers