feat: s2s add high dep tokens (#809)
Goal is to compute high dep tokens
https://app.shortcut.com/galileo/story/9232/dq-save-nessec-column-for-high-dep-tokens?vc_group_by=day&ct_workflow=all&cf_workflow=500000005

The DQ work requires saving the target tokens. We later groupby them
(and the ID is not enough as we need the tokens for the UI).

Decoding the target token ids is fast enough in my opinion. Here it's
taking ~10 sec for 10K rows and ~1M tokens
![Screenshot 2023-11-28 at 10 17
41 AM](https://github.com/rungalileo/dataquality/assets/112427971/e71256c8-7d5b-4456-bdcf-ff58ba54d650).

---------

Co-authored-by: Elliott <elliott@rungalileo.io>