Clean up data `formaters`, removing need for passing `Seq2SeqDataLogger` (#810)
***PR Description***

One of the current design choices for the `data_logger` formatters, was
to pass the actual `Seq2SeqDataLogger` to the `.format()` function. This
was nice for dev speed, but is not the nicest for three primary reason
(that I see):
1. It can cause annoying trickiness around avoiding circular imports.
For example, I am working on the generation PR for Decoder-Only models,
and I want to use `data_formatters`; however, when I try to do so, I get
circular import errors.
2. Overall it seems like not the greatest python "style".
3. It is not as necessary as we maybe initially thought.

To elaborate on point (3), you will see in this PR that we basically
need to address two main issues when removing the dependance on having
the actual `Seq2SeqDataLogger`:
1. Saving information to the `LoggerConfig`: Accomplished by passing in
two additional data fields to the `format` function: a) `ids` and b)
`split_key`
2. Assigning offsets and positions: Accomplished by returning the
"formatted" `AlignedData` object.

With these two fairly simple extensions we are able to simplify the
`format(.)` function!

Note: One addition that we have to make is extending the
`AlignedTokenData` class to have an `append(.)` function. This is
necessary because for DecoderOnly models we process the aligned token
data one sample at a time. The function we add is simple and follows
this structure:
```
# The AlignedTokenData object that we are building up sample by sample
aligned_data: AlignedTokenData

# Aligned token data for a *single* sample - i.e. batch dim == 1
sample_aligned_token_data: AlignedTokenData

# This will append the `offset` and `position` data from the sample
# to the internal fields of `aligned_data`.
aligned_data.append(sample_aligned_token_data)
```